async_openai
Module types

Types used in OpenAI API requests and responses. These types are created from component schemas in the OpenAPI spec

Structs

AssistantFileObject

A list of Files attached to an assistant.

AssistantObject

Represents an assistant that can call the model and use tools.

AssistantToolsCode

Code interpreter tool

AssistantToolsFunction

Function tool

AssistantToolsRetrieval

Retrieval tool

AudioInput

Category

CategoryScore

A list of the categories along with their scores as predicted by model.

ChatChoice

ChatCompletionFunctions

ChatCompletionFunctionsArgs

Builder for ChatCompletionFunctions.

ChatCompletionMessageToolCall

ChatCompletionMessageToolCallChunk

ChatCompletionNamedToolChoice

Specifies a tool the model should use. Use to force the model to call a specific function.

ChatCompletionRequestAssistantMessage

ChatCompletionRequestAssistantMessageArgs

Builder for ChatCompletionRequestAssistantMessage.

ChatCompletionRequestFunctionMessage

ChatCompletionRequestFunctionMessageArgs

Builder for ChatCompletionRequestFunctionMessage.

ChatCompletionRequestMessageContentPartImage

ChatCompletionRequestMessageContentPartImageArgs

Builder for ChatCompletionRequestMessageContentPartImage.

ChatCompletionRequestMessageContentPartText

ChatCompletionRequestMessageContentPartTextArgs

Builder for ChatCompletionRequestMessageContentPartText.

ChatCompletionRequestSystemMessage

ChatCompletionRequestSystemMessageArgs

Builder for ChatCompletionRequestSystemMessage.

ChatCompletionRequestToolMessage

Tool message

ChatCompletionRequestToolMessageArgs

Builder for ChatCompletionRequestToolMessage.

ChatCompletionRequestUserMessage

ChatCompletionRequestUserMessageArgs

Builder for ChatCompletionRequestUserMessage.

ChatCompletionResponseFormat

ChatCompletionResponseMessage

A chat completion message generated by the model.

ChatCompletionResponseStreamMessage

ChatCompletionStreamResponseDelta

A chat completion delta generated by streamed model responses.

ChatCompletionTool

ChatCompletionToolArgs

Builder for ChatCompletionTool.

Choice

CodeInterpreter

CompletionUsage

Usage statistics for the completion request.

ContentModerationResult

CreateAssistantFileRequest

CreateAssistantRequest

CreateAssistantRequestArgs

Builder for CreateAssistantRequest.

CreateChatCompletionRequest

CreateChatCompletionRequestArgs

Builder for CreateChatCompletionRequest.

CreateChatCompletionResponse

Represents a chat completion response returned by model, based on the provided input.

CreateChatCompletionStreamResponse

Represents a streamed chunk of a chat completion response returned by model, based on the provided input.

CreateCompletionRequest

CreateCompletionRequestArgs

Builder for CreateCompletionRequest.

CreateCompletionResponse

CreateEditRequest

CreateEditRequestArgs

Builder for CreateEditRequest.

CreateEditResponse

CreateEmbeddingRequest

CreateEmbeddingRequestArgs

Builder for CreateEmbeddingRequest.

CreateEmbeddingResponse

CreateFileRequest

CreateFileRequestArgs

Builder for CreateFileRequest.

CreateFineTuneRequest

CreateFineTuneRequestArgs

Builder for CreateFineTuneRequest.

CreateFineTuningJobRequest

CreateFineTuningJobRequestArgs

Builder for CreateFineTuningJobRequest.

CreateImageEditRequest

CreateImageEditRequestArgs

Builder for CreateImageEditRequest.

CreateImageRequest

CreateImageRequestArgs

Builder for CreateImageRequest.

CreateImageVariationRequest

CreateImageVariationRequestArgs

Builder for CreateImageVariationRequest.

CreateMessageRequest

CreateMessageRequestArgs

Builder for CreateMessageRequest.

CreateModerationRequest

CreateModerationRequestArgs

Builder for CreateModerationRequest.

CreateModerationResponse

Represents policy compliance report by OpenAI’s content moderation model against a given input.

CreateRunRequest

CreateRunRequestArgs

Builder for CreateRunRequest.

CreateSpeechRequest

CreateSpeechRequestArgs

Builder for CreateSpeechRequest.

CreateSpeechResponse

CreateThreadAndRunRequest

CreateThreadAndRunRequestArgs

Builder for CreateThreadAndRunRequest.

CreateThreadRequest

CreateThreadRequestArgs

Builder for CreateThreadRequest.

CreateTranscriptionRequest

CreateTranscriptionRequestArgs

Builder for CreateTranscriptionRequest.

CreateTranscriptionResponse

CreateTranslationRequest

CreateTranslationRequestArgs

Builder for CreateTranslationRequest.

CreateTranslationResponse

DeleteAssistantFileResponse

Deletes the association between the assistant and the file, but does not delete the File object itself.

DeleteAssistantResponse

DeleteFileResponse

DeleteMessageResponse

DeleteModelResponse

DeleteThreadResponse

Embedding

Represents an embedding vector returned by embedding endpoint.

EmbeddingUsage

FileCitation

FileInput

FilePath

FineTune

FineTuneEvent

FineTuneJobError

For fine-tuning jobs that have failed, this will contain more information on the cause of the failure.

FineTuningJob

The fine_tuning.job object represents a fine-tuning job that has been created through the API.

FineTuningJobEvent

Fine-tuning job event object

FunctionCall

The name and arguments of a function that should be called, as generated by the model.

FunctionCallStream

FunctionName

FunctionObject

Hyperparameters

ImageFile

ImageInput

ImageUrl

ImageUrlArgs

Builder for ImageUrl.

ImagesResponse

LastError

ListAssistantFilesResponse

ListAssistantsResponse

ListFilesResponse

ListFineTuneEventsResponse

ListFineTuneEventsStreamResponse

ListFineTuneResponse

ListFineTuningJobEventsResponse

ListMessageFilesResponse

ListMessagesResponse

ListModelResponse

ListPaginatedFineTuningJobsResponse

ListRunStepsResponse

ListRunsResponse

Logprobs

MessageContentImageFileObject

References an image File in the content of a message.

MessageContentTextAnnotationsFileCitationObject

MessageContentTextAnnotationsFilePathObject

MessageContentTextObject

The text content that is part of a message.

MessageCreation

MessageFileObject

A list of files attached to a message.

MessageObject

Represents a message within a thread.

Model

Describes an OpenAI model offering that can be used with the API.

ModifyAssistantRequest

ModifyAssistantRequestArgs

Builder for ModifyAssistantRequest.

ModifyMessageRequest

ModifyRunRequest

ModifyThreadRequest

OpenAIFile

The File object represents a document that has been uploaded to OpenAI.

RequiredAction

RunObject

Represents an execution run on a thread.

RunStepDetailsMessageCreationObject

Details of the message creation by the run step.

RunStepDetailsToolCallsCodeObject

Code interpreter tool call

RunStepDetailsToolCallsCodeOutputImageObject

RunStepDetailsToolCallsCodeOutputLogsObject

Text output from the Code Interpreter tool call as part of a run step.

RunStepDetailsToolCallsFunctionObject

RunStepDetailsToolCallsObject

Details of the tool call.

RunStepDetailsToolCallsRetrievalObject

RunStepObject

Represents a step in execution of a run.

RunToolCallObject

SubmitToolOutputs

SubmitToolOutputsRunRequest

TextData

ThreadObject

Represents a thread that contains messages.

ToolsOutputs

ToolsOutputsArgs

Builder for ToolsOutputs.

Enums

AssistantTools

AudioResponseFormat

ChatCompletionFunctionCall

ChatCompletionRequestMessage

ChatCompletionRequestMessageContentPart

ChatCompletionRequestUserMessageContent

ChatCompletionResponseFormatType

ChatCompletionToolChoiceOption

Controls which (if any) function is called by the model. none means the model will not call a function and instead generates a message. auto means the model can pick between generating a message or calling a function. Specifying a particular function via {"type: "function", "function": {"name": "my_function"}} forces the model to call that function. none is the default when no functions are present. auto is the default if functions are present.

ChatCompletionToolType

CodeInterpreterOutput

CompletionFinishReason

DallE2ImageSize

EmbeddingInput

EncodingFormat

FineTuningJobStatus

FinishReason

Image

ImageModel

ImageQuality

ImageSize

ImageStyle

ImageUrlDetail

LastErrorCode

Level

MessageContent

MessageContentTextAnnotations

MessageRole

ModerationInput

NEpochs

OpenAIFilePurpose

Prompt

ResponseFormat

Role

RunStatus

RunStepDetailsToolCalls

RunStepType

SpeechModel

SpeechResponseFormat

StepDetails

Stop

TextModerationModel

Voice

Type Aliases

ChatCompletionResponseStream

Parsed server side events stream until an [DONE] is received from server.

CompletionResponseStream

Parsed server side events stream until an [DONE] is received from server.

FineTuneEventsResponseStream

Parsed server side events stream until an [DONE] is received from server.