async_openai::types
Struct Category

pub struct Category {
    pub hate: bool,
    pub hate_threatening: bool,
    pub harassment: bool,
    pub harassment_threatening: bool,
    pub self_harm: bool,
    pub self_harm_intent: bool,
    pub self_harm_instructions: bool,
    pub sexual: bool,
    pub sexual_minors: bool,
    pub violence: bool,
    pub violence_graphic: bool,
}

Fields
hate: bool

Content that expresses, incites, or promotes hate based on race, gender, ethnicity, religion, nationality, sexual orientation, disability status, or caste. Hateful content aimed at non-protected groups (e.g., chess players) is harrassment.

hate_threatening: bool

Hateful content that also includes violence or serious harm towards the targeted group based on race, gender, ethnicity, religion, nationality, sexual orientation, disability status, or caste.

harassment: bool

Content that expresses, incites, or promotes harassing language towards any target.

harassment_threatening: bool

Harassment content that also includes violence or serious harm towards any target.

self_harm: bool

Content that promotes, encourages, or depicts acts of self-harm, such as suicide, cutting, and eating disorders.

self_harm_intent: bool

Content where the speaker expresses that they are engaging or intend to engage in acts of self-harm, such as suicide, cutting, and eating disorders.

self_harm_instructions: bool

Content that encourages performing acts of self-harm, such as suicide, cutting, and eating disorders, or that gives instructions or advice on how to commit such acts.

sexual: bool

Content meant to arouse sexual excitement, such as the description of sexual activity, or that promotes sexual services (excluding sex education and wellness).

sexual_minors: bool

Sexual content that includes an individual who is under 18 years old.

violence: bool

Content that depicts death, violence, or physical injury.

violence_graphic: bool

Content that depicts death, violence, or physical injury in graphic detail.