async_openai
Module types

Types used in OpenAI API requests and responses. These types are created from component schemas in the OpenAPI spec

Structs

AudioInput

Category

CategoryScore

ChatChoice

ChatCompletionFunctions

ChatCompletionFunctionsArgs

Builder for ChatCompletionFunctions.

ChatCompletionRequestMessage

ChatCompletionRequestMessageArgs

Builder for ChatCompletionRequestMessage.

ChatCompletionResponseMessage

ChatCompletionResponseStreamMessage

ChatCompletionStreamResponseDelta

Choice

ContentModerationResult

CreateChatCompletionRequest

CreateChatCompletionRequestArgs

Builder for CreateChatCompletionRequest.

CreateChatCompletionResponse

CreateChatCompletionStreamResponse

CreateCompletionRequest

CreateCompletionRequestArgs

Builder for CreateCompletionRequest.

CreateCompletionResponse

CreateEditRequest

CreateEditRequestArgs

Builder for CreateEditRequest.

CreateEditResponse

CreateEmbeddingRequest

CreateEmbeddingRequestArgs

Builder for CreateEmbeddingRequest.

CreateEmbeddingResponse

CreateFileRequest

CreateFileRequestArgs

Builder for CreateFileRequest.

CreateFineTuneRequest

CreateFineTuneRequestArgs

Builder for CreateFineTuneRequest.

CreateImageEditRequest

CreateImageEditRequestArgs

Builder for CreateImageEditRequest.

CreateImageRequest

CreateImageRequestArgs

Builder for CreateImageRequest.

CreateImageVariationRequest

CreateImageVariationRequestArgs

Builder for CreateImageVariationRequest.

CreateModerationRequest

CreateModerationRequestArgs

Builder for CreateModerationRequest.

CreateModerationResponse

CreateTranscriptionRequest

CreateTranscriptionRequestArgs

Builder for CreateTranscriptionRequest.

CreateTranscriptionResponse

CreateTranslationRequest

CreateTranslationRequestArgs

Builder for CreateTranslationRequest.

CreateTranslationResponse

DeleteFileResponse

DeleteModelResponse

Embedding

EmbeddingUsage

FileInput

FineTune

FineTuneEvent

FunctionCall

The name and arguments of a function that should be called, as generated by the model.

FunctionCallStream

ImageInput

ImageResponse

ListFilesResponse

ListFineTuneEventsResponse

ListFineTuneEventsStreamResponse

ListFineTuneResponse

ListModelResponse

Logprobs

Model

OpenAIFile

Usage

Enums

AudioResponseFormat

ChatCompletionFunctionCall

EmbeddingInput

ImageData

ImageSize

ModerationInput

Prompt

ResponseFormat

Role

Stop

TextModerationModel

Type Aliases

ChatCompletionResponseStream

Parsed server side events stream until an [DONE] is received from server.

CompletionResponseStream

Parsed server side events stream until an [DONE] is received from server.

FineTuneEventsResponseStream

Parsed server side events stream until an [DONE] is received from server.